Cleanup warnings - work in progress.

+++ firmware/ggl/ggl/ggl_opmask.c
@@ -9,7 +9,7 @@ unsigned int ggl_opmask(unsigned int dest,unsigned int src, int tcol)
     register unsigned int res  =0;
     for(f=0;f<8;++f,src>>=4,dest>>=4)
     {
-        if((src&0xf)==tcol) res|=dest&0xf;
+        if((src&0xf)==(unsigned int)tcol) res|=dest&0xf;
         else res|=src&0xf;
         res= (res>>4) | (res<<28);
 
@@ -30,7 +30,7 @@ unsigned int ggl_opmaskcol(unsigned int dest,unsigned int src, int tcol, int new
     register unsigned int res  =0;
     for(f=0;f<8;++f,src>>=4,dest>>=4)
     {
-        if((src&0xf)==tcol) res|=dest&0xf;
+        if((src&0xf)==(unsigned int)tcol) res|=dest&0xf;
         else res|=src&newcolor;
         res= (res>>4) | (res<<28);
     }
